fix: resolve all clippy warnings across workspace
- dns-server: collapse nested ifs, drop redundant pattern match, allow too_many_arguments on write_soa
- seeder-core: rename Config::from_str to parse, iterator-based pchMessageStart loop, drop needless borrows
- seeder-core: Default impl for AddrStat, drop identity op, canonical PartialOrd delegating to Ord
- seeder-core: move lookup_host before test module, drop op_ref, allow too_many_arguments on serialize_version_payload, drop redundant cast
- seeder: use existing IpCallback alias, replace manual map

+++ crates/seeder-core/src/p2p/message.rs
@@ -157,6 +157,7 @@ impl Address {
     }
 }
 
+#[allow(clippy::too_many_arguments)]
 pub fn serialize_version_payload(
     proto_version: u32,
     services: u64,
@@ -171,7 +172,7 @@ pub fn serialize_version_payload(
     let mut payload = Vec::new();
     payload.extend_from_slice(&proto_version.to_le_bytes());
     payload.extend_from_slice(&services.to_le_bytes());
-    payload.extend_from_slice(&(timestamp as i64).to_le_bytes());
+    payload.extend_from_slice(&timestamp.to_le_bytes());
     payload.extend_from_slice(&addr_recv.serialize());
     payload.extend_from_slice(&addr_from.serialize());
     payload.extend_from_slice(&nonce.to_le_bytes());
